Changed matrix.h. Added GPU matrix multiply.

+++ GpuComputing/include/gpu_kernel.h
@@ -10,8 +10,6 @@
 
 #endif
 
-#include "cuda_runtime.h"
-#include "device_launch_parameters.h"
 #include "matrix.h"
 
 #ifdef __cplusplus
@@ -20,7 +18,7 @@ extern "C"
 #endif
 
 GPU_API_EXPORT void GPU_test(PerfComparison::Matrix<float> const& A, PerfComparison::Matrix<float> const& B, PerfComparison::Matrix<float> &C, float& ComputingTime);
-GPU_API_EXPORT void DisplayMatrix(float const*, size_t,char*);
+GPU_API_EXPORT void DisplayMatrix(double const*, size_t,char*);
 	
 #ifdef __cplusplus
 }
